fix(build): k_cache_line_size hardcoded to 64, dropped the redundant static_cast<int32_t> around std::stoi(v), ASan job now tees configure/build/test output to log files and uploads them as an asan-logs artifact on failure (14-day retention)

+++ src/rest/client.cpp
@@ -37,7 +37,7 @@ struct Client::Impl {
     void update_rate_limits(const detail::HttpResponse& resp) noexcept {
         auto try_parse = [](const std::string& v) noexcept -> int32_t {
             try {
-                return static_cast<int32_t>(std::stoi(v));
+                return std::stoi(v);
             } catch (const std::exception&) {
                 return -1;
             }
